+++ src/main/java/ar/edu/unlam/nuralign/application/ports/in/medicationTracker/FindMedicationTrackerByEffectiveDateUseCase.java
@@ -4,6 +4,6 @@ import ar.edu.unlam.nuralign.domain.models.MedicationTracker;
 
 import java.util.List;
 
-public interface FindAllMedicationTrackersByEffectiveDateUseCase {
-    List<MedicationTracker> findAllMedicationTrackersByEffectiveDate(String effectiveDate);
+public interface FindMedicationTrackerByEffectiveDateUseCase {
+    MedicationTracker findMedicationTrackerByEffectiveDate(String effectiveDate);
 }