+++ src/main/java/ar/edu/unlam/nuralign/application/ports/out/SleepTrackerRepositoryPort.java
@@ -10,5 +10,6 @@ public interface SleepTrackerRepositoryPort {
     Optional<SleepTracker> findById(Long id);
     List<SleepTracker> findAll();
     List<SleepTracker> findAllByPatientId(Long patientId);
-    List<SleepTracker> findAllSleepTrackersByPatientIdAndEffectiveDate(Long patientId, String effectiveDate);
+    SleepTracker findSleepTrackerByPatientIdAndEffectiveDate(Long patientId, String effectiveDate);
+    Optional<SleepTracker> update(SleepTracker sleepTracker, Long patientId, String effectiveDate);
 }