+++ src/main/java/ar/edu/unlam/nuralign/infrastructure/adapters/JpaMedicationTrackerRepositoryAdapter.java
@@ -2,12 +2,14 @@ package ar.edu.unlam.nuralign.infrastructure.adapters;
 
 import ar.edu.unlam.nuralign.application.ports.out.MedicationTrackerRepositoryPort;
 import ar.edu.unlam.nuralign.domain.models.MedicationTracker;
+import ar.edu.unlam.nuralign.infrastructure.entities.MedicationTrackerEntity;
 import ar.edu.unlam.nuralign.infrastructure.mappers.MedicationTrackerMapper;
 import ar.edu.unlam.nuralign.infrastructure.repositories.JpaMedicationTrackerRepository;
 import org.springframework.stereotype.Component;
 
 import java.time.LocalDate;
 import java.util.List;
+import java.util.Optional;
 
 @Component
 public class JpaMedicationTrackerRepositoryAdapter implements MedicationTrackerRepositoryPort {
@@ -38,26 +40,33 @@ public class JpaMedicationTrackerRepositoryAdapter implements MedicationTrackerR
 
     @Override
     public List<MedicationTracker> findAllByPatientId(Long patientId) {
-        return jpaMedicationTrackerRepository.findAllByPatientId(patientId)
-                .stream()
+        return jpaMedicationTrackerRepository.findAllByPatientId(patientId).stream()
                 .map(MedicationTrackerMapper::toModel)
                 .toList();
     }
 
     @Override
-    public List<MedicationTracker> findAllByEffectiveDate(LocalDate effectiveDate) {
-        return jpaMedicationTrackerRepository.findAllByEffectiveDate(effectiveDate)
-                .stream()
-                .map(MedicationTrackerMapper::toModel)
-                .toList();
+    public MedicationTracker findByEffectiveDate(LocalDate effectiveDate) {
+        return MedicationTrackerMapper.toModel(jpaMedicationTrackerRepository.findByEffectiveDate(effectiveDate));
     }
 
     @Override
-    public List<MedicationTracker> findAllByPatientIdAndEffectiveDate(Long patientId, LocalDate effectiveDate) {
-        return jpaMedicationTrackerRepository.findAllByPatientIdAndEffectiveDate(patientId, effectiveDate)
-                .stream()
-                .map(MedicationTrackerMapper::toModel)
-                .toList();
+    public MedicationTracker findByPatientIdAndEffectiveDate(Long patientId, LocalDate effectiveDate) {
+        return MedicationTrackerMapper.toModel(jpaMedicationTrackerRepository.findByPatientIdAndEffectiveDate(patientId, effectiveDate));
+    }
+
+    @Override
+    public Optional<MedicationTracker> update(MedicationTracker medicationTracker, Long patientId, LocalDate effectiveDate) {
+        if (!jpaMedicationTrackerRepository.existsByPatientIdAndEffectiveDate(patientId, effectiveDate)) {
+            return Optional.empty();
+        }
+        MedicationTrackerEntity entityToUpdate = jpaMedicationTrackerRepository.findByPatientIdAndEffectiveDate(
+                patientId, effectiveDate);
+        if (medicationTracker.getTakenFlag() != null) {
+            entityToUpdate.setTakenFlag(medicationTracker.getTakenFlag());
+        }
+
+        return Optional.of(MedicationTrackerMapper.toModel(jpaMedicationTrackerRepository.save(entityToUpdate)));
     }
 
 }