+++ src/main/java/ar/edu/unlam/nuralign/infrastructure/config/ApplicationConfig.java
@@ -1,18 +1,14 @@
 package ar.edu.unlam.nuralign.infrastructure.config;
 
-import ar.edu.unlam.nuralign.application.ports.in.patientTherapist.AssignPatientToTherapistUseCase;
 import ar.edu.unlam.nuralign.application.ports.out.*;
 import ar.edu.unlam.nuralign.application.services.*;
 import ar.edu.unlam.nuralign.application.usecases.medication.*;
 import ar.edu.unlam.nuralign.application.usecases.medicationTracker.*;
-import ar.edu.unlam.nuralign.application.usecases.moodTracker.CreateMoodTrackerUseCaseImpl;
-import ar.edu.unlam.nuralign.application.usecases.moodTracker.FindAllMoodTrackersByPatientIdAndEffectiveDateUseCaseImpl;
-import ar.edu.unlam.nuralign.application.usecases.moodTracker.FindAllMoodTrackersUseCaseImpl;
-import ar.edu.unlam.nuralign.application.usecases.moodTracker.FindMoodTrackerUseCaseImpl;
+import ar.edu.unlam.nuralign.application.usecases.moodTracker.*;
 import ar.edu.unlam.nuralign.application.usecases.patient.*;
-import ar.edu.unlam.nuralign.application.usecases.patientTherapist.AssignPatientToTherapistUseCaseImpl;
-import ar.edu.unlam.nuralign.application.usecases.patientTherapist.DeletePatientTherapistAssignationUseCaseImpl;
-import ar.edu.unlam.nuralign.application.usecases.patientTherapist.FindAllPatientTherapistAssignationsUseCaseImpl;
+//import ar.edu.unlam.nuralign.application.usecases.patientTherapist.AssignPatientToTherapistUseCaseImpl;
+//import ar.edu.unlam.nuralign.application.usecases.patientTherapist.DeletePatientTherapistAssignationUseCaseImpl;
+//import ar.edu.unlam.nuralign.application.usecases.patientTherapist.FindAllPatientTherapistAssignationsUseCaseImpl;
 import ar.edu.unlam.nuralign.application.usecases.sleepTracker.*;
 import ar.edu.unlam.nuralign.application.usecases.therapist.*;
 import ar.edu.unlam.nuralign.application.usecases.therapySession.CreateTherapySessionUseCaseImpl;
@@ -29,14 +25,14 @@ import org.springframework.context.annotation.Configuration;
 @Configuration
 public class ApplicationConfig {
 
-    @Bean
-    public PatientTherapistService patientTherapistService(PatientTherapistRepositoryPort patientTherapistRepositoryPort) {
-        return new PatientTherapistService(
-                new AssignPatientToTherapistUseCaseImpl(patientTherapistRepositoryPort),
-                new FindAllPatientTherapistAssignationsUseCaseImpl(patientTherapistRepositoryPort),
-                new DeletePatientTherapistAssignationUseCaseImpl(patientTherapistRepositoryPort)
-        );
-    }
+//    @Bean
+//    public PatientTherapistService patientTherapistService(PatientTherapistRepositoryPort patientTherapistRepositoryPort) {
+//        return new PatientTherapistService(
+//                new AssignPatientToTherapistUseCaseImpl(patientTherapistRepositoryPort),
+//                new FindAllPatientTherapistAssignationsUseCaseImpl(patientTherapistRepositoryPort),
+//                new DeletePatientTherapistAssignationUseCaseImpl(patientTherapistRepositoryPort)
+//        );
+//    }
 
     @Bean
     public TherapistService therapistService(TherapistRepositoryPort therapistRepositoryPort) {
@@ -55,19 +51,21 @@ public class ApplicationConfig {
                 new CreateMoodTrackerUseCaseImpl(moodTrackerRepositoryPort),
                 new FindMoodTrackerUseCaseImpl(moodTrackerRepositoryPort),
                 new FindAllMoodTrackersUseCaseImpl(moodTrackerRepositoryPort),
-                new FindAllMoodTrackersByPatientIdAndEffectiveDateUseCaseImpl(moodTrackerRepositoryPort)
+                new FindMoodTrackerByPatientIdAndEffectiveDateUseCaseImpl(moodTrackerRepositoryPort),
+                new UpdateMoodTrackerUseCaseImpl(moodTrackerRepositoryPort)
         );
     }
 
     @Bean
-    public MedicationTrackerService medicationTrackerService(MedicationTrackerRepositoryPort medicationTrackerRepositoryPort) {
-        return new MedicationTrackerService(
+    public MedicationTrackersService medicationTrackerService(MedicationTrackerRepositoryPort medicationTrackerRepositoryPort) {
+        return new MedicationTrackersService(
                 new FindMedicationTrackerUseCaseImpl(medicationTrackerRepositoryPort),
-                new FindAllMedicationTrackersByEffectiveDateUseCaseImpl(medicationTrackerRepositoryPort),
+                new FindMedicationTrackerByEffectiveDateUseCaseImpl(medicationTrackerRepositoryPort),
                 new FindAllMedicationTrackersByPatientIdUseCaseImpl(medicationTrackerRepositoryPort),
                 new FindAllMedicationTrackersUseCaseImpl(medicationTrackerRepositoryPort),
                 new CreateMedicationTrackerUseCaseImpl(medicationTrackerRepositoryPort),
-                new FindAllMedicationTrackerByPatientIdAndEffectiveDateUseCaseImpl(medicationTrackerRepositoryPort)
+                new FindMedicationTrackerByPatientIdAndEffectiveDateUseCaseImpl(medicationTrackerRepositoryPort),
+                new UpdateMedicationTrackerUseCaseImpl(medicationTrackerRepositoryPort)
         );
     }
 
@@ -101,7 +99,8 @@ public class ApplicationConfig {
                 new FindSleepTrackerUseCaseImpl(sleepTrackerRepositoryPort),
                 new FindAllSleepTrackersUseCaseImpl(sleepTrackerRepositoryPort),
                 new FindAllSleepTrackersByPatientIdImpl(sleepTrackerRepositoryPort),
-                new FindAllSleepTrackersByPatientIdAndEffectiveDateImpl(sleepTrackerRepositoryPort)
+                new FindSleepTrackerByPatientIdAndEffectiveDateImpl(sleepTrackerRepositoryPort),
+                new UpdateSleepTrackerUseCaseImpl(sleepTrackerRepositoryPort)
         );
     }
 
@@ -122,10 +121,10 @@ public class ApplicationConfig {
         return jpaTherapySessionAdapter;
     }
 
-    @Bean
-    public PatientTherapistRepositoryPort patientTherapistRepositoryPort(JpaPatientTherapistRepositoryAdapter jpaPatientTherapistRepositoryAdapter) {
-        return jpaPatientTherapistRepositoryAdapter;
-    }
+//    @Bean
+//    public PatientTherapistRepositoryPort patientTherapistRepositoryPort(JpaPatientTherapistRepositoryAdapter jpaPatientTherapistRepositoryAdapter) {
+//        return jpaPatientTherapistRepositoryAdapter;
+//    }
 
     @Bean
     public JpaTherapySessionAdapter therapySessionAdapter(JpaTherapySessionRepository adapter)
@@ -192,10 +191,10 @@ public JpaTurnTherapistRepositoryAdapter turnTherapistRepositoryAdapter(JpaTurnT
         return jpaMedicationTrackerRepositoryAdapter;
     }
 
-    @Bean
-    public JpaPatientTherapistRepositoryAdapter jpaPatientTherapistRepositoryAdapter(JpaPatientTherapistRepository adapter) {
-        return new JpaPatientTherapistRepositoryAdapter(adapter);
-    }
+//    @Bean
+//    public JpaPatientTherapistRepositoryAdapter jpaPatientTherapistRepositoryAdapter(JpaPatientTherapistRepository adapter) {
+//        return new JpaPatientTherapistRepositoryAdapter(adapter);
+//    }
 
     @Bean
     public JpaTherapistRepositoryAdapter jpaTherapistRepositoryAdapter(JpaTherapistRepository adapter) {